[players] Move annotation pickers to players/bars/
ColorPicker, PencilPicker and ShapePicker are only consumed by
PlayerAnnotationBar — they belong next to it under players/bars/
rather than as generic widgets/.
