+++ src/components/players/bars/ColorPicker.vue
