+++ src/components/players/bars/PencilPicker.vue
