+++ src/components/players/bars/ShapePicker.vue
